Function now in correct file
