+++ pages/api/case-note/[id].ts
@@ -2,9 +2,12 @@ import StatusCodes from 'http-status-codes';
 import { finishSubmission, patchSubmissionForStep } from 'lib/submissions';
 import { FormikValues } from 'formik';
 import { AxiosError } from 'axios';
-import { apiHandler, AuthenticatedNextApiHandler } from 'lib/apiHandler';
+import {
+  apiHandler,
+  AuthenticatedNextApiHandler,
+  handleAxiosError,
+} from 'lib/apiHandler';
 import { middleware as csrfMiddleware } from 'lib/csrfToken';
-import { handleAxiosError } from 'lib/errorHandler';
 
 const handler: AuthenticatedNextApiHandler = async (
   req,
