+++ pages/api/cases/[caseId].ts
@@ -3,8 +3,11 @@ import { StatusCodes } from 'http-status-codes';
 import { getCase } from 'lib/cases';
 import { middleware as csrfMiddleware } from 'lib/csrfToken';
 import { AxiosError } from 'axios';
-import { apiHandler, AuthenticatedNextApiHandler } from 'lib/apiHandler';
-import { handleAxiosError } from 'lib/errorHandler';
+import {
+  apiHandler,
+  AuthenticatedNextApiHandler,
+  handleAxiosError,
+} from 'lib/apiHandler';
 
 const endpoint: AuthenticatedNextApiHandler = async (req, res) => {
   const { caseId, residentId, ...params } = req.query;
