+++ pages/api/cases/index.ts
@@ -2,10 +2,12 @@ import { StatusCodes } from 'http-status-codes';
 
 import { getCases, addCase } from 'lib/cases';
 import { middleware as csrfMiddleware } from 'lib/csrfToken';
-import { handleAxiosError } from 'lib/errorHandler';
-
 import { AxiosError } from 'axios';
-import { apiHandler, AuthenticatedNextApiHandler } from 'lib/apiHandler';
+import {
+  apiHandler,
+  AuthenticatedNextApiHandler,
+  handleAxiosError,
+} from 'lib/apiHandler';
 
 const endpoint: AuthenticatedNextApiHandler = async (req, res) => {
   switch (req.method) {
