+++ pages/api/mash-referral/reset.ts
@@ -3,8 +3,12 @@ import { middleware as csrfMiddleware } from 'lib/csrfToken';
 
 import { AxiosError } from 'axios';
 import { resetDummyData } from 'lib/mashReferral';
-import { apiHandler, AuthenticatedNextApiHandler } from 'lib/apiHandler';
-import { handleAxiosError } from 'lib/errorHandler';
+import {
+  apiHandler,
+  AuthenticatedNextApiHandler,
+  handleAxiosError,
+} from 'lib/apiHandler';
+
 const endpoint: AuthenticatedNextApiHandler = async (req, res) => {
   switch (req.method) {
     case 'POST':
