+++ pages/api/residents/[id]/allocations/[allocationId].ts
@@ -3,8 +3,11 @@ import { StatusCodes } from 'http-status-codes';
 import { getResidentAllocation } from 'lib/allocatedWorkers';
 import { middleware as csrfMiddleware } from 'lib/csrfToken';
 import { AxiosError } from 'axios';
-import { apiHandler, AuthenticatedNextApiHandler } from 'lib/apiHandler';
-import { handleAxiosError } from 'lib/errorHandler';
+import {
+  apiHandler,
+  AuthenticatedNextApiHandler,
+  handleAxiosError,
+} from 'lib/apiHandler';
 
 const endpoint: AuthenticatedNextApiHandler = async (req, res) => {
   switch (req.method) {
