+++ pages/api/submissions/[id]/delete.ts
@@ -1,9 +1,12 @@
 import StatusCodes from 'http-status-codes';
 import { deleteSubmission } from 'lib/submissions';
 import { AxiosError } from 'axios';
-import { apiHandler, AuthenticatedNextApiHandler } from 'lib/apiHandler';
+import {
+  apiHandler,
+  AuthenticatedNextApiHandler,
+  handleAxiosError,
+} from 'lib/apiHandler';
 import { middleware as csrfMiddleware } from 'lib/csrfToken';
-import { handleAxiosError } from 'lib/errorHandler';
 
 const handler: AuthenticatedNextApiHandler = async (
   req,
