+++ pages/api/workers/[id]/allocations.ts
@@ -3,10 +3,13 @@ import { StatusCodes } from 'http-status-codes';
 import { getWorker } from 'lib/workers';
 import { getAllocationsByWorker } from 'lib/allocatedWorkers';
 import { middleware as csrfMiddleware } from 'lib/csrfToken';
-import { handleAxiosError } from 'lib/errorHandler';
 
 import { AxiosError } from 'axios';
-import { apiHandler, AuthenticatedNextApiHandler } from 'lib/apiHandler';
+import {
+  apiHandler,
+  AuthenticatedNextApiHandler,
+  handleAxiosError,
+} from 'lib/apiHandler';
 
 const endpoint: AuthenticatedNextApiHandler = async (req, res) => {
   switch (req.method) {
